Added Loading animation when no Internet or Loading
