+++ Bruin Bite/Helpers/NetworkManager.swift
@@ -1,0 +1,82 @@
+//
+//  NetworkManager.swift
+//  Bruin Bite
+//
+//  Created by Sahen Rai on 11/27/18.
+//  Copyright © 2018 Dont Eat Alone. All rights reserved.
+//
+
+import Foundation
+import Reachability
+
+class NetworkManager: NSObject {
+    
+    var reachability: Reachability!
+    
+    // Create a singleton instance
+    static let sharedInstance: NetworkManager = { return NetworkManager() }()
+    
+    
+    override init() {
+        super.init()
+        
+        // Initialise reachability
+        reachability = Reachability()!
+        
+        // Register an observer for the network status
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(networkStatusChanged(_:)),
+            name: .reachabilityChanged,
+            object: reachability
+        )
+        
+        do {
+            // Start the network status notifier
+            try reachability.startNotifier()
+        } catch {
+            print("Unable to start notifier")
+        }
+    }
+    
+    @objc func networkStatusChanged(_ notification: Notification) {
+        // Do something globally here!
+    }
+    
+    static func stopNotifier() -> Void {
+        do {
+            // Stop the network status notifier
+            try (NetworkManager.sharedInstance.reachability).startNotifier()
+        } catch {
+            print("Error stopping notifier")
+        }
+    }
+    
+    // Network is reachable
+    static func isReachable(completed: @escaping (NetworkManager) -> Void) {
+        if (NetworkManager.sharedInstance.reachability).connection != .none {
+            completed(NetworkManager.sharedInstance)
+        }
+    }
+    
+    // Network is unreachable
+    static func isUnreachable(completed: @escaping (NetworkManager) -> Void) {
+        if (NetworkManager.sharedInstance.reachability).connection == .none {
+            completed(NetworkManager.sharedInstance)
+        }
+    }
+    
+    // Network is reachable via WWAN/Cellular
+    static func isReachableViaWWAN(completed: @escaping (NetworkManager) -> Void) {
+        if (NetworkManager.sharedInstance.reachability).connection == .cellular {
+            completed(NetworkManager.sharedInstance)
+        }
+    }
+    
+    // Network is reachable via WiFi
+    static func isReachableViaWiFi(completed: @escaping (NetworkManager) -> Void) {
+        if (NetworkManager.sharedInstance.reachability).connection == .wifi {
+            completed(NetworkManager.sharedInstance)
+        }
+    }
+}
